Local de gravity-gun atualizado e README com dicas
