+++ README.md
@@ -1,4 +1,6 @@
 # montagem_de_peca
 Trabalho do 3° semestre com realidade virtual de Laboratório 3 com o Professor Giuliano
 ## setup
-npm run build
+1° npm install -g browserify
+2° npm install
+3° npm run build
