+++ view/modeling/index.html
@@ -23,6 +23,6 @@
         </a-camera>
         
     </a-scene>
-    <script src="bundle.js"></script>
+    <script src="../../bundle.js"></script>
 </body>
 </html>
